+++ src/array/methods.js
@@ -13,5 +13,7 @@ export { default as min } from './min';
 export { default as median } from './median';
 export { default as none } from './none';
 export { default as remove } from './remove';
+export { default as sample } from './sample';
+export { default as shuffle } from './shuffle';
 export { default as some } from './some';
 export { default as sum } from './sum';